requirements.txt
Flask
oic
cherrypy
pyyaml
yubico-client==1.9.1
PyJWT
